Add request packet steering in target

+++ initiator.c
@@ -147,7 +147,8 @@ ethblk_initiator_cmd_dump_to_string(struct ethblk_initiator_cmd *cmd, char *ptr,
 		       cmd->id, cmd, req, cmd->t, cmd->hctx_idx, cmd->gen_id,
 		       cmd->retries, cmd->d->name, req_op(req), req_name);
 	if (has_lba) {
-		snprintf(ptr + ret, n - ret, " lba %llu len %u", blk_rq_pos(req),
+		snprintf(ptr + ret, n - ret, " lba %llu len %u",
+			 (unsigned long long)blk_rq_pos(req),
 			 blk_rq_bytes(req));
 	}
 }
@@ -1147,8 +1148,8 @@ ethblk_initiator_cmd_rw(struct ethblk_initiator_cmd *cmd, bool last)
 			"disk %s cmd[%d] req %p req_op %d lba %llu len %u "
 			"retries %d requested > max payload (%d)\n",
 			cmd->d->name, cmd->id, req, req_op(req),
-			blk_rq_pos(req), blk_rq_bytes(req), cmd->retries,
-			cmd->d->max_payload);
+			(unsigned long long)blk_rq_pos(req), blk_rq_bytes(req),
+			cmd->retries, cmd->d->max_payload);
 		status = BLK_STS_NOTSUPP;
 		goto out;
 	}
@@ -1249,12 +1250,14 @@ ethblk_initiator_blk_queue_request(struct blk_mq_hw_ctx *hctx,
 	if (cmd->retries)
 		dprintk(info,
 			"cmd[%d] req %p req_op %d lba %llu len %u retries %d\n",
-			cmd->id, bd->rq, req_op(bd->rq), blk_rq_pos(bd->rq),
+			cmd->id, bd->rq, req_op(bd->rq),
+			(unsigned long long)blk_rq_pos(bd->rq),
 			blk_rq_bytes(bd->rq), cmd->retries);
 	else
 		dprintk(debug,
 			"cmd[%d] req %p req_op %d lba %llu len %u retries %d\n",
-			cmd->id, bd->rq, req_op(bd->rq), blk_rq_pos(bd->rq),
+			cmd->id, bd->rq, req_op(bd->rq),
+			(unsigned long long)blk_rq_pos(bd->rq),
 			blk_rq_bytes(bd->rq), cmd->retries);
 	cmd->retries = 0;
 	cmd->time_queued = cmd->time_requeued = current_time = ktime_get_ns();
